Refactor: RecordService Reader/Updater 분리
